% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{download_mpinterests}
\alias{download_mpinterests}
\title{Download Register of Members' Financial Interests}
\usage{
download_mpinterests(contents_page)
}
\arguments{
\item{contents_page}{A character string of the contents page URL
for an edition of the Register of Members' Financial Interests.}
}
\description{
\code{download_mpinterests} scrapes entries in the Register of Members'
Financial Interests for each MP. There is no need to assign when calling
{download_mpinterests}, although if you do the returned objected is a raw
list of the entries.
}
