% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_donations_and_support}
\alias{get_donations_and_support}
\title{Create a tibble of "Category 2: Donations and other support" entries.}
\usage{
get_donations_and_support()
}
\value{
A tibble of "Category 2: Donations and other support" entries.
}
\description{
\code{get_donations_and_support} gets entries under "Category 2: Donations
and other support" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
