% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_downloaded_mpinterests}
\alias{get_downloaded_mpinterests}
\title{Get the cached Register of Members' Financial Interests}
\usage{
get_downloaded_mpinterests(get = c("interests", "date"))
}
\value{
The cached Register of Members' Financial Interests as a list
}
\description{
\code{get_cache} gets the downloaded Register of Members' Financial
Interests which is used in all get_* functions.
}
\keyword{internal}
