% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_employment_and_earnings}
\alias{get_employment_and_earnings}
\title{Create a tibble of "Category 1: Employment and earnings" entries.}
\usage{
get_employment_and_earnings()
}
\value{
A tibble of "Category 1: Employment and earnings" entries.
}
\description{
\code{get_employment_and_earnings} gets entries under "Category 1:
Employment and earnings" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
