% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_family_employed}
\alias{get_family_employed}
\title{Create a tibble of "Category 9: Family members employed" entries.}
\usage{
get_family_employed()
}
\value{
A tibble of "Category 9: Family members employed" entries.
}
\description{
\code{get_family_employed} gets entries under "Category 9:
Family members employed" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
