% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_family_lobbying}
\alias{get_family_lobbying}
\title{Create a tibble of "Category 10: Family members lobbying" entries.}
\usage{
get_family_lobbying()
}
\value{
A tibble of "Category 10: Family members lobbying" entries.
}
\description{
\code{get_family_lobbying} gets entries under "Category 10:
Family members lobbying" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
