% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_gifts_and_benefits_uk}
\alias{get_gifts_and_benefits_uk}
\title{Create a tibble of "Category 3: Gifts and benefits from UK sources" entries.}
\usage{
get_gifts_and_benefits_uk()
}
\value{
A tibble of "Category 3: Gifts and benefits from UK sources" entries.
}
\description{
\code{get_gifts_and_benefits_uk} gets entries under "Category 3: Gifts and
benefits from UK sources" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
