% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_land_and_property}
\alias{get_land_and_property}
\title{Create a tibble of "Category 6: Land and property" entries.}
\usage{
get_land_and_property()
}
\value{
A tibble of "Category 6: Land and property" entries.
}
\description{
\code{get_land_and_property} gets entries under "Category 6:
Land and property" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
