% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_miscellaneous}
\alias{get_miscellaneous}
\title{Create a tibble of "Category 8: Miscellaneous" entries.}
\usage{
get_miscellaneous()
}
\value{
A tibble of "Category 8: Miscellaneous" entries.
}
\description{
\code{get_miscellaneous} gets entries under "Category 8:
Miscellaneous" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
