% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_shareholdings}
\alias{get_shareholdings}
\title{Create a tibble of "Category 7: Shareholdings" entries.}
\usage{
get_shareholdings()
}
\value{
A tibble of "Category 7: Shareholdings" entries.
}
\description{
\code{get_shareholdings} gets entries under "Category 7:
Shareholdings" in the Register of Members' Financial Interests
and returns a tibble. Each row is one entry.
}
