% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_visits_outside_uk}
\alias{get_visits_outside_uk}
\title{Create a tibble of "Category 4: Visits outside the UK" entries.}
\usage{
get_visits_outside_uk()
}
\value{
A tibble of "Category 4: Visits outside the UK" entries.
}
\description{
\code{get_visits_outside_uk} gets entries under "Category 4: Visits outside
the UK" in the Register of Members' Financial Interests and returns a tibble.
Each row is one entry.
}
