% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_category_entries}
\alias{handle_category_entries}
\title{Hangle category entries}
\usage{
handle_category_entries(contents, category, end_pattern)
}
\description{
Hangle category entries
}
\keyword{internal}
