% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle.R
\name{handle_category_tibble}
\alias{handle_category_tibble}
\title{Handle category tibble}
\usage{
handle_category_tibble(contents, category, end_pattern, date)
}
\description{
Handle category tibble
}
\keyword{internal}
